requirements_backend.txt
# ViralSpot Backend API Requirements
# =================================

# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and storage
supabase==2.0.4
postgrest==0.12.0
gotrue==1.3.4
realtime==1.0.4
storage3==0.7.4
httpx==0.25.2

# Data processing
pydantic==2.5.1
python-multipart==0.0.6

# Environment and utilities
python-dotenv==1.0.0

# Existing pipeline dependencies (if you want to run the full pipeline)
requests==2.31.0
aiohttp==3.9.1
openai==1.3.9

# Optional: Development tools
pytest==7.4.3
pytest-asyncio==0.21.1